+++ src/settings.js
@@ -1160,8 +1160,6 @@ var LINKABLE = false;
 // codebase builds nicely in forward compatible manner.
 // Changes enabled by this:
 //
-//   - The C define EMSCRIPTEN is not defined (__EMSCRIPTEN__ always is, and
-//     is the correct thing to use).
 //   - STRICT_JS is enabled.
 //   - IGNORE_MISSING_MAIN is disabled.
 //   - AUTO_JS_LIBRARIES is disabled.